// src/generated/arrow/flight/Flight_pb.ts
export type FlightDescriptor = Message<"arrow.flight.protocol.FlightDescriptor"> & {
  /**
   * @generated from field: arrow.flight.protocol.FlightDescriptor.DescriptorType type = 1;
   */
  type: FlightDescriptor_DescriptorType;

  /**
   *
   * Opaque value used to express a command. Should only be defined when
   * type = CMD.
   *
   * @generated from field: bytes cmd = 2;
   */
  cmd: Uint8Array;

  /**
   *
   * List of strings identifying a particular dataset. Should only be defined
   * when type = PATH.
   *
   * @generated from field: repeated string path = 3;
   */
  path: string[];
};

/**
 *
 * The name or tag for a Flight. May be used as a way to retrieve or generate
 * a flight or be used to expose a set of previously defined flights.
 *
 * @generated from message arrow.flight.protocol.FlightDescriptor
 */
export type FlightDescriptorJson = {
  /**
   * @generated from field: arrow.flight.protocol.FlightDescriptor.DescriptorType type = 1;
   */
  type?: FlightDescriptor_DescriptorTypeJson;

  /**
   *
   * Opaque value used to express a command. Should only be defined when
   * type = CMD.
   *
   * @generated from field: bytes cmd = 2;
   */
  cmd?: string;

  /**
   *
   * List of strings identifying a particular dataset. Should only be defined
   * when type = PATH.
   *
   * @generated from field: repeated string path = 3;
   */
  path?: string[];
};

/**
 * Describes the message arrow.flight.protocol.FlightDescriptor.
 * Use `create(FlightDescriptorSchema)` to create a new message.
 */
export const FlightDescriptorSchema: GenMessage<FlightDescriptor, {jsonType: FlightDescriptorJson}> = /*@__PURE__*/
  messageDesc(file_arrow_flight_Flight, 9);

/**
 *
 * Describes what type of descriptor is defined.
 *
 * @generated from enum arrow.flight.protocol.FlightDescriptor.DescriptorType
 */
export enum FlightDescriptor_DescriptorType {
  /**
   * Protobuf pattern, not used.
   *
   * @generated from enum value: UNKNOWN = 0;
   */
  UNKNOWN = 0,

  /**
   *
   * A named path that identifies a dataset. A path is composed of a string
   * or list of strings describing a particular dataset. This is conceptually
   *  similar to a path inside a filesystem.
   *
   * @generated from enum value: PATH = 1;
   */
  PATH = 1,

  /**
   *
   * An opaque command to generate a dataset.
   *
   * @generated from enum value: CMD = 2;
   */
  CMD = 2,
}

/**
 *
 * Describes what type of descriptor is defined.
 *
 * @generated from enum arrow.flight.protocol.FlightDescriptor.DescriptorType
 */
export type FlightDescriptor_DescriptorTypeJson = "UNKNOWN" | "PATH" | "CMD";
// ...
export type FlightData = Message<"arrow.flight.protocol.FlightData"> & {
  /**
   *
   * The descriptor of the data. This is only relevant when a client is
   * starting a new DoPut stream.
   *
   * @generated from field: arrow.flight.protocol.FlightDescriptor flight_descriptor = 1;
   */
  flightDescriptor?: FlightDescriptor;

  /**
   *
   * Header for message data as described in Message.fbs::Message.
   *
   * @generated from field: bytes data_header = 2;
   */
  dataHeader: Uint8Array;

  /**
   *
   * Application-defined metadata.
   *
   * @generated from field: bytes app_metadata = 3;
   */
  appMetadata: Uint8Array;

  /**
   *
   * The actual batch of Arrow data. Preferably handled with minimal-copies
   * coming last in the definition to help with sidecar patterns (it is
   * expected that some implementations will fetch this field off the wire
   * with specialized code to avoid extra memory copies).
   *
   * @generated from field: bytes data_body = 1000;
   */
  dataBody: Uint8Array;
};

/**
 *
 * A batch of Arrow data as part of a stream of batches.
 *
 * @generated from message arrow.flight.protocol.FlightData
 */
export type FlightDataJson = {
  /**
   *
   * The descriptor of the data. This is only relevant when a client is
   * starting a new DoPut stream.
   *
   * @generated from field: arrow.flight.protocol.FlightDescriptor flight_descriptor = 1;
   */
  flightDescriptor?: FlightDescriptorJson;

  /**
   *
   * Header for message data as described in Message.fbs::Message.
   *
   * @generated from field: bytes data_header = 2;
   */
  dataHeader?: string;

  /**
   *
   * Application-defined metadata.
   *
   * @generated from field: bytes app_metadata = 3;
   */
  appMetadata?: string;

  /**
   *
   * The actual batch of Arrow data. Preferably handled with minimal-copies
   * coming last in the definition to help with sidecar patterns (it is
   * expected that some implementations will fetch this field off the wire
   * with specialized code to avoid extra memory copies).
   *
   * @generated from field: bytes data_body = 1000;
   */
  dataBody?: string;
};

/**
 * Describes the message arrow.flight.protocol.FlightData.
 * Use `create(FlightDataSchema)` to create a new message.
 */
export const FlightDataSchema: GenMessage<FlightData, {jsonType: FlightDataJson}> = /*@__PURE__*/
  messageDesc(file_arrow_flight_Flight, 18);